Move OpenGL controllers to Castaway.Level
