+++ Castaway.Level/Controllers/ShaderController.cs
@@ -2,17 +2,15 @@ using System;
 using System.Linq;
 using Castaway.Assets;
 using Castaway.Base;
-using Castaway.Level;
 using Castaway.Rendering;
 using Castaway.Rendering.Lighting;
 using Castaway.Rendering.Objects;
 using Castaway.Rendering.Shaders;
 using Serilog;
 
-namespace Castaway.OpenGL.Controllers;
+namespace Castaway.Level.Controllers;
 
 [ControllerName("Shader")]
-[Imports(typeof(OpenGLImpl))]
 public class ShaderController : Controller
 {
 	private static readonly ILogger Logger = CastawayGlobal.GetLogger();
