--- Castaway.OpenGL.Controllers/Castaway.OpenGL.Controllers.csproj
@@ -1,29 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <Title>Castaway Levels (OpenGL)</Title>
-        <PackageId>Castaway.OpenGL.Controllers</PackageId>
-
-        <!-- Common -->
-        <PackageVersion>1.0.0-pre1</PackageVersion>
-        <AssemblyVersion>1.0.0.1</AssemblyVersion>
-        <FileVersion>2022.1</FileVersion>
-        <TargetFramework>net6.0</TargetFramework>
-        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-        <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-        <Authors>Liam Cole</Authors>
-        <RepositoryUrl>https://github.com/LiamCoalStudio/castaway</RepositoryUrl>
-        <RepositoryType>GitHub</RepositoryType>
-        <Nullable>enable</Nullable>
-        <Copyright>[c] LiamCoalStudio 2021</Copyright>
-        <PackageProjectUrl>https://github.com/LiamCoalStudio/castaway</PackageProjectUrl>
-        <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\Castaway.Base\Castaway.Base.csproj"/>
-        <ProjectReference Include="..\Castaway.Level\Castaway.Level.csproj"/>
-        <ProjectReference Include="..\Castaway.OpenGL\Castaway.OpenGL.csproj"/>
-    </ItemGroup>
-
-</Project>
