+++ Castaway.sln
@@ -30,8 +30,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Castaway.Input", "Castaway.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Castaway.UI", "Castaway.UI\Castaway.UI.csproj", "{7573D949-9070-45AD-8E58-FE081E6BFAB4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Castaway.OpenGL.Controllers", "Castaway.OpenGL.Controllers\Castaway.OpenGL.Controllers.csproj", "{405BE028-E34B-4C78-8DFD-8A0CDD9CF904}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -90,10 +88,6 @@ Global
 		{7573D949-9070-45AD-8E58-FE081E6BFAB4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7573D949-9070-45AD-8E58-FE081E6BFAB4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7573D949-9070-45AD-8E58-FE081E6BFAB4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{405BE028-E34B-4C78-8DFD-8A0CDD9CF904}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{405BE028-E34B-4C78-8DFD-8A0CDD9CF904}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{405BE028-E34B-4C78-8DFD-8A0CDD9CF904}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{405BE028-E34B-4C78-8DFD-8A0CDD9CF904}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{0D12B6EE-9A58-46D1-8E9B-AF17DA3256E8} = {18F0C133-7D5D-4E3A-B8E2-A190D0DF80F3}
@@ -107,6 +101,5 @@ Global
 		{99715B59-81CB-4A71-B510-49343A8A943F} = {18F0C133-7D5D-4E3A-B8E2-A190D0DF80F3}
 		{78EE8151-FDC3-47A9-BABA-CEB634628F47} = {18F0C133-7D5D-4E3A-B8E2-A190D0DF80F3}
 		{7573D949-9070-45AD-8E58-FE081E6BFAB4} = {18F0C133-7D5D-4E3A-B8E2-A190D0DF80F3}
-		{405BE028-E34B-4C78-8DFD-8A0CDD9CF904} = {18F0C133-7D5D-4E3A-B8E2-A190D0DF80F3}
 	EndGlobalSection
 EndGlobal
