+++ Test/Program.cs
@@ -3,8 +3,8 @@ using Castaway.Assets;
 using Castaway.Base;
 using Castaway.Input;
 using Castaway.Level;
+using Castaway.Level.Controllers;
 using Castaway.OpenGL;
-using Castaway.OpenGL.Controllers;
 using Castaway.Rendering;
 using OpenTK.Windowing.GraphicsLibraryFramework;
 using Window = Castaway.Rendering.Window;
